requirement.txt
1.certifi==2020.6.20
2.chardet==3.0.4
3.click==7.1.2
4.Flask==1.1.2
5.idna==2.10
6.itsdangerous==1.1.0
7.Jinja2==2.11.2
8.joblib==0.15.1
9.jsonify==0.5
10.MarkupSafe==1.1.1
11.numpy==1.19.0
12.requests==2.24.0
13.scikit-learn==0.23.1
14.scipy==1.5.0
15.sklearn==0.0
16.threadpoolctl==2.1.0
17.urllib3==1.25.9
18.Werkzeug==1.0.1
19.wincertstore==0.2
20.gunicorn
21.Footer